[Addition] update dev's parameter when in config function
If configuration function is public, dev's parameter should be update, not just provide a spi transmit interface. (et. When we use nrf24l01 connet multi slave or something else, Tx address must be changed. So nrf24l01 LL function is important to export to user. In condition like that, maybe the dev's parameter should be changed)